/feature_route_details_images fix models, remove annotation creator to helper class, remove center coordinates calculation

+++ app/src/main/java/com/example/gh_coursework/ui/helper/EventHelper.kt
@@ -3,16 +3,16 @@ package com.example.gh_coursework.ui.helper
 import android.os.SystemClock
 import android.view.MotionEvent
 
-fun createOnMapClickEvent(center: Pair<Float, Float>): Pair<MotionEvent, MotionEvent> {
+fun createOnMapClickEvent(center: Pair<Int, Int>): Pair<MotionEvent, MotionEvent> {
     val downTime = SystemClock.uptimeMillis()
     val eventTime = SystemClock.uptimeMillis() + 10
     val downAction = MotionEvent.obtain(
         downTime, eventTime, MotionEvent.ACTION_DOWN,
-        center.first , center.second, 0
+        center.first.toFloat(), center.second.toFloat(), 0
     )
     val upAction = MotionEvent.obtain(
         downTime, eventTime, MotionEvent.ACTION_UP,
-        center.first , center.second, 0
+        center.first.toFloat(), center.second.toFloat(), 0
     )
 
     return Pair(downAction, upAction)
